tratando erros

+++ src/modules/user/useCases/deleteUser/deleteUserUseCase.ts
@@ -1,3 +1,4 @@
+import { AppError } from "../../../../erros/appError";
 import { User } from "../../model/user";
 import { IPostgreSQLDBRepository } from "../../repositories/IPostgreSQLDBRepository";
 
@@ -6,6 +7,15 @@ class DeleteUserUseCase {
     constructor(private userRepository: IPostgreSQLDBRepository){}
 
     async execute(user: User){
+
+        const {name } = user 
+
+        const userExists = await this.userRepository.findByName(name);
+
+        if(!userExists){
+            throw new AppError("User don't exists!")
+        }
+
         await this.userRepository.deleteUser(user)
     }
 }
